@@ -23,10 +23,12 @@
     "@types/jstimezonedetect": "1.0.0",
     "@types/lodash": "0.0.27",
     "@types/moment-timezone": "0.2.34",
-    "@types/node": "7.0.3",
+    "@types/node": "7.0.8",
     "@types/oclazyload": "0.0.31",
     "@types/stripe": "0.0.34",
     "@types/ua-parser-js": "0.7.30",
+    "@types/qs": "6.2.31",
+    "@types/express": "4.0.35",
     "angular": "1.5.8",
     "angular-animate": "1.5.8",
     "angular-hammer": "2.2.0",
@@ -35,19 +37,20 @@
     "angular-sanitize": "1.5.8",
     "angular-ui-router": "1.0.0-rc.1",
     "angular-ui-tree": "2.22.5",
+    "axios": "0.15.3",
     "bootstrap-stylus": "hworld/bootstrap-stylus#master",
     "chart.js": "1.0.2",
-    "codemirror": "5.24.2",
+    "codemirror": "5.25.0",
     "core-js": "2.4.1",
-    "date-fns": "1.28.0",
+    "date-fns": "1.28.1",
     "decompress-zip": "0.3.0",
     "favico.js": "0.3.10",
-    "game-jolt-frontend-lib": "gamejolt/frontend-lib#master",
+    "game-jolt-frontend-lib": "gamejolt/frontend-lib#more-rewrites",
     "hammerjs": "2.0.8",
-    "jquery": "3.1.1",
+    "jquery": "3.2.1",
     "jquery-jcrop": "0.9.13",
     "jstimezonedetect": "1.0.6",
-    "moment": "2.17.1",
+    "moment": "2.18.1",
     "moment-timezone": "0.5.11",
     "ng-file-upload": "12.2.13",
     "ng-metadata": "4.0.1",
@@ -61,12 +64,19 @@
     "tc-angular-chartjs": "1.0.11",
     "ua-parser-js": "0.7.12",
     "v-tooltip": "1.1.6",
-    "vue": "2.1.10",
-    "vue-class-component": "4.4.0",
+    "vue": "2.2.5",
+    "vue-class-component": "5.0.0",
     "vue-gettext": "gamejolt/vue-gettext#hworld-lazyload",
-    "vue-property-decorator": "3.4.0",
+    "vue-property-decorator": "4.0.0",
     "vue-shortkey": "2.1.0",
-    "vue-touch": "2.0.0-beta.4"
+    "vue-touch": "2.0.0-beta.4",
+    "vuex": "2.2.1",
+    "vue-router": "2.3.0",
+    "vuex-class": "0.1.3",
+    "vee-validate": "2.0.0-beta.25",
+    "vue-server-renderer": "2.2.5",
+    "express": "4.15.2",
+    "vue-ssr-webpack-plugin": "1.0.2"
   },
   "optionalDependencies": {
     "appdmg": "^0.3.5"
