@@ -1,3 +1,4 @@
+import Vue from 'vue';
 import { ActivityFeedItem, ActivityFeedInput } from './item-service';
 import { FiresidePost } from '../../../../lib/gj-lib-client/components/fireside/post/post-model';
 
@@ -6,14 +7,18 @@ export class ActivityFeedContainer
 	items: ActivityFeedItem[] = [];
 	games: { [k: string]: any } = {};
 
-	hasItems = false;
 	viewedItems: string[] = [];
 	expandedItems: string[] = [];
-	notificationWatermark?: number;  // Timestamp.
+	notificationWatermark = 0;  // Timestamp.
 	reachedEnd = false;
 
-	private _activeItem: ActivityFeedItem | undefined;
+	private _activeItem: ActivityFeedItem | null = null;
 	private _scroll = 0;
+
+	get hasItems()
+	{
+		return this.items.length > 0;
+	}
 
 	constructor( items: ActivityFeedInput[], notificationWatermark?: number )
 	{
@@ -28,7 +33,6 @@ export class ActivityFeedContainer
 	{
 		const items = _items.map( item => new ActivityFeedItem( item ) );
 		this.items = items.concat( this.items );
-		this.hasItems = this.items.length > 0;
 		this._processGames();
 	}
 
@@ -36,22 +40,19 @@ export class ActivityFeedContainer
 	{
 		const items = _items.map( item => new ActivityFeedItem( item ) );
 		this.items = this.items.concat( items );
-		this.hasItems = this.items.length > 0;
 		this._processGames();
 	}
 
 	update( _item: ActivityFeedInput )
 	{
 		const item = new ActivityFeedItem( _item );
-		const index = _.findIndex( this.items, {
-			type: item.type,
-			feedItem: {
-				id: item.feedItem.id,
-			}
+		const index = this.items.findIndex( ( i ) =>
+		{
+			return i.type === item.type && i.feedItem.id === item.feedItem.id;
 		} );
 
 		if ( index >= 0 ) {
-			this.items[ index ] = item;
+			Vue.set( this.items, index, item );
 		}
 
 		this._processGames();
@@ -60,17 +61,17 @@ export class ActivityFeedContainer
 	remove( _item: ActivityFeedInput )
 	{
 		const item = new ActivityFeedItem( _item );
-		_.remove( this.items, {
-			type: item.type,
-			feedItem: {
-				id: item.feedItem.id,
-			}
+		const index = this.items.findIndex( ( i ) =>
+		{
+			return i.type === item.type && i.feedItem.id === item.feedItem.id;
 		} );
-		this.hasItems = this.items.length > 0;
+
+		this.items.splice( index, 1 );
+
 		this._processGames();
 	}
 
-	setActive( active: ActivityFeedItem | undefined )
+	setActive( active: ActivityFeedItem | null )
 	{
 		this._activeItem = active;
 	}
@@ -92,13 +93,13 @@ export class ActivityFeedContainer
 
 	viewed( item: ActivityFeedItem )
 	{
-		if ( this.viewedItems.indexOf( item.id ) != -1 ) {
+		if ( this.viewedItems.indexOf( item.id ) !== -1 ) {
 			return;
 		}
 
 		this.viewedItems.push( item.id );
 
-		if ( item.type == 'devlog-post' ) {
+		if ( item.type === 'devlog-post' ) {
 			const feedItem = <FiresidePost>item.feedItem;
 			feedItem.$viewed();
 		}
@@ -106,13 +107,13 @@ export class ActivityFeedContainer
 
 	expanded( item: ActivityFeedItem )
 	{
-		if ( this.expandedItems.indexOf( item.id ) != -1 ) {
+		if ( this.expandedItems.indexOf( item.id ) !== -1 ) {
 			return;
 		}
 
 		this.expandedItems.push( item.id );
 
-		if ( item.type == 'devlog-post' ) {
+		if ( item.type === 'devlog-post' ) {
 			const feedItem = <FiresidePost>item.feedItem;
 			feedItem.$expanded();
 		}
@@ -128,10 +129,10 @@ export class ActivityFeedContainer
 		for ( const item of this.items ) {
 			if ( item.feedItem instanceof FiresidePost ) {
 				if ( !this.games[ item.feedItem.game.id ] ) {
-					this.games[ item.feedItem.game.id ] = item.feedItem.game;
+					Vue.set( this.games, item.feedItem.game.id, item.feedItem.game );
 				}
 				else {
-					item.feedItem.game = this.games[ item.feedItem.game.id ];
+					Vue.set( item.feedItem, 'game', this.games[ item.feedItem.game.id ] );
 				}
 			}
 		}
