@@ -1,6 +1,21 @@
-import { provide } from 'ng-metadata/core';
-import { ThumbnailComponent } from './thumbnail-directive';
+import Vue from 'vue';
+import { Component, Prop } from 'vue-property-decorator';
+import * as View from '!view!./thumbnail.html?style=./thumbnail.styl';
 
-export default angular.module( 'App.Channel.Thumbnail', [] )
-.directive( ...provide( ThumbnailComponent ) )
-.name;
+import { Channels } from '../channels-service';
+import { AppTrackEvent } from '../../../../lib/gj-lib-client/components/analytics/track-event.directive.vue';
+
+@View
+@Component({
+	name: 'channel-thumbnail',
+	directives: {
+		AppTrackEvent,
+	}
+})
+export class AppChannelThumbnail extends Vue
+{
+	@Prop( String ) channel: string;
+	@Prop( Number ) gamesCount?: number;
+
+	Channels = Channels;
+}
