@@ -1,7 +1,7 @@
-angular.module( 'App.Game.Filtering', [] );
+// angular.module( 'App.Game.Filtering', [] );
 
-require( './filtering-service' );
-require( './input-directive' );
-require( './sidebar-widget-directive' );
-require( './tags-directive' );
-require( './widget-directive' );
+// require( './filtering-service' );
+// require( './input-directive' );
+// require( './sidebar-widget-directive' );
+// require( './tags-directive' );
+// require( './widget-directive' );
