@@ -1,3 +1,3 @@
-angular.module( 'App.Score.Overview', [ 'App.Score.List', 'App.Score.Feed', 'App.Score.ScoreboardPopover', 'gj.Game.ScoreTable', 'gj.User.GameScore' ] );
+angular.module( 'App.Score.Overview', [] );
 
 require( './overview-directive' );
