@@ -1,213 +1,216 @@
-import { Injectable, Inject } from 'ng-metadata/core';
-
+// import * as angular from 'angular';
 import { Screen } from '../../../lib/gj-lib-client/components/screen/screen-service';
-import { BroadcastModal } from '../broadcast-modal/broadcast-modal.service';
-import { getProvider, lazyload } from '../../../lib/gj-lib-client/utils/utils';
-import { App } from '../../app-service';
+// import { BroadcastModal } from '../broadcast-modal/broadcast-modal.service';
+// import { App } from '../../app-service';
 import { Api } from '../../../lib/gj-lib-client/components/api/api.service';
 import { Settings } from '../settings/settings.service';
+import { Chat } from '../chat/chat.service';
+import { GameCollection } from '../game/collection/collection.model';
 
-@Injectable( 'Shell' )
+let bootstrapPromiseResolve: any = undefined;
+
 export class Shell
 {
-	collections: any[];
-	followedCollection?: any;
-	developerCollection?: any;
-	ownedCollection?: any;
-	recommendedCollection?: any;
-	bundleCollections: any[];
+	// static notificationCount = 0;
+	// static collections: GameCollection[] = [];
+	// static followedCollection?: GameCollection = undefined;
+	// static developerCollection?: GameCollection = undefined;
+	// static ownedCollection?: GameCollection = undefined;
+	// static recommendedCollection?: GameCollection = undefined;
+	// static bundleCollections: GameCollection[] = [];
 
-	isBootstrapped = false;
-	bootstrapPromise: Promise<void>;
-	private bootstrapPromiseResolve: ng.IQResolveReject<void>;
+	// static isBootstrapped = false;
+	// static bootstrapPromise = new Promise<void>(
+	// 	( resolve ) => bootstrapPromiseResolve = resolve
+	// );
 
-	private _isLeftPaneSticky = true;
-	private _isLeftPaneOverlayed = false;
-	private _isRightPaneOverlayed = false;
+	// private static _isLeftPaneSticky = true;
+	// private static _isLeftPaneOverlayed = false;
+	// private static _isRightPaneOverlayed = false;
 
-	private backdrop?: any;
-	private chat?: any;
+	private static backdrop?: any = undefined;
+	private static chat?: typeof Chat = undefined;
 
-	constructor(
-		@Inject( '$rootScope' ) $rootScope: ng.IRootScopeService,
-		@Inject( 'Backdrop' ) private Backdrop: any,
-		@Inject( 'GameCollection' ) private collectionModel: any,
-		@Inject( 'BroadcastModal' ) private broadcastModal: BroadcastModal,
-		@Inject( 'hotkeys' ) private hotkeys: ng.hotkeys.HotkeysProvider,
-	)
+	static init()
 	{
-		this.bootstrapPromise = new Promise<void>( ( resolve ) => this.bootstrapPromiseResolve = resolve );
+		// const $rootScope = getProvider<ng.IRootScopeService>( '$rootScope' );
+		// const hotkeys = getProvider<ng.hotkeys.HotkeysProvider>( 'hotkeys' );
 
-		this._isLeftPaneSticky = Settings.get( 'sidebar' );
+		// this._isLeftPaneSticky = Settings.get( 'sidebar' );
 
-		/**
-		 * After changing states, hide all overlays.
-		 */
-		$rootScope.$on( '$stateChangeStart', () =>
-		{
-			// When navigating, only change the overlayed states.
-			this._isLeftPaneOverlayed = false;
-			this._isRightPaneOverlayed = false;
-			this.checkBackdrop();
-		} );
+		// /**
+		//  * After changing states, hide all overlays.
+		//  */
+		// $rootScope.$on( '$stateChangeStart', () =>
+		// {
+		// 	// When navigating, only change the overlayed states.
+		// 	this._isLeftPaneOverlayed = false;
+		// 	this._isRightPaneOverlayed = false;
+		// 	this.checkBackdrop();
+		// } );
 
-		hotkeys.add( {
-			combo: 'm',
-			description: 'Toggle the sidebar menu.',
-			callback: () => this.toggleLeftPane(),
-		} );
+		// hotkeys.add( {
+		// 	combo: 'm',
+		// 	description: 'Toggle the sidebar menu.',
+		// 	callback: () => this.toggleLeftPane(),
+		// } );
 
-		$rootScope.$watch( () => !!getProvider<App>( 'App' ).user, ( isLoggedIn ) =>
-		{
-			if ( isLoggedIn ) {
-				this.bootstrap();
-			}
-			else {
-				this.clear();
-			}
+		// $rootScope.$watch( () => !!getProvider<App>( 'App' ).user, async ( isLoggedIn ) =>
+		// {
+		// 	if ( isLoggedIn ) {
+		// 		this.bootstrap();
+		// 	}
+		// 	else {
+		// 		this.clear();
+		// 	}
 
-			if ( isLoggedIn ) {
-				require.ensure( [], () => lazyload( () =>
-				{
-					require( '../chat/chat.module' );
-				} ), 'chat' )
-				.then( () =>
-				{
-					// Connect to chat.
-					this.chat = getProvider<any>( 'Chat' );
-					this.chat.connect();
+		// 	if ( isLoggedIn ) {
+		// 		const mod = await $import( '../chat/chat.service' );
 
-					hotkeys.add( {
-						combo: 'c',
-						description: 'Toggle the chat.',
-						callback: () => this.toggleRightPane(),
-					} );
-				} );
-			}
-			else {
-				// Log out of chat. This will notify other tabs to disconnect from the server too.
-				if ( this.chat ) {
-					if ( this.chat.client ) {
-						this.chat.client.logOut();
-					}
-					this.hotkeys.del( 'c' );
-				}
+		// 		this.chat = mod.Chat as typeof Chat;
+		// 		this.chat.connect();
 
-				this.chat = undefined;
-			}
-		} );
+		// 		hotkeys.add( {
+		// 			combo: 'c',
+		// 			description: 'Toggle the chat.',
+		// 			callback: () => this.toggleRightPane(),
+		// 		} );
+		// 	}
+		// 	else {
+		// 		// Log out of chat. This will notify other tabs to disconnect from the server too.
+		// 		if ( this.chat ) {
+		// 			if ( this.chat.client ) {
+		// 				this.chat.client.logout();
+		// 			}
+		// 			hotkeys.del( 'c' );
+		// 		}
+
+		// 		this.chat = undefined;
+		// 	}
+		// } );
 	}
 
-	get isLeftPaneVisible()
-	{
-		if ( Screen.isDesktop ) {
-			return this._isLeftPaneSticky;
-		}
+	// static get isLeftPaneVisible()
+	// {
+	// 	if ( Screen.isDesktop ) {
+	// 		return this._isLeftPaneSticky;
+	// 	}
 
-		return this._isLeftPaneOverlayed;
-	}
+	// 	return this._isLeftPaneOverlayed;
+	// }
 
-	get isRightPaneVisible()
-	{
-		return this._isRightPaneOverlayed;
-	}
+	// static get isRightPaneVisible()
+	// {
+	// 	return this._isRightPaneOverlayed;
+	// }
 
-	toggleLeftPane()
-	{
-		if ( Screen.isDesktop ) {
-			this._isLeftPaneSticky = !this._isLeftPaneSticky;
-		}
-		else {
-			this._isLeftPaneOverlayed = !this._isLeftPaneOverlayed;
-		}
+	// static toggleLeftPane()
+	// {
+	// 	if ( Screen.isDesktop ) {
+	// 		this._isLeftPaneSticky = !this._isLeftPaneSticky;
+	// 	}
+	// 	else {
+	// 		this._isLeftPaneOverlayed = !this._isLeftPaneOverlayed;
+	// 	}
 
-		this._isRightPaneOverlayed = false;
-		this.checkBackdrop();
-		Settings.set( 'sidebar', this._isLeftPaneSticky );
+	// 	this._isRightPaneOverlayed = false;
+	// 	this.checkBackdrop();
+	// 	Settings.set( 'sidebar', this._isLeftPaneSticky );
 
-		return this;
-	}
+	// 	return this;
+	// }
 
-	toggleRightPane()
-	{
-		this._isRightPaneOverlayed = !this._isRightPaneOverlayed;
-		this._isLeftPaneOverlayed = false;
-		this.checkBackdrop();
+	// static toggleRightPane()
+	// {
+	// 	this._isRightPaneOverlayed = !this._isRightPaneOverlayed;
+	// 	this._isLeftPaneOverlayed = false;
+	// 	this.checkBackdrop();
 
-		return this;
-	}
+	// 	return this;
+	// }
 
-	private shouldShowLeftPaneBackdrop()
-	{
-		return this._isLeftPaneOverlayed && Screen.isMobile;
-	}
+	// static addPlaylist( playlist: GameCollection )
+	// {
+	// 	this.collections.push( playlist );
+	// }
 
-	private checkBackdrop()
-	{
-		// Ensure we have a backdrop if anything is overlayed.
-		// Otherwise ensure the backdrop is gone.
-		if ( this._isRightPaneOverlayed || this.shouldShowLeftPaneBackdrop() ) {
-			if ( this.backdrop ) {
-				return;
-			}
+	// static removePlaylist( playlist: GameCollection )
+	// {
+	// 	const index = this.collections.findIndex( ( i ) => i._id === playlist._id );
+	// 	this.collections.splice( index, 1 );
+	// }
 
-			this.backdrop = new this.Backdrop(
-				angular.element( document.getElementById( 'shell-body' ) as HTMLElement )
-			);
+	// private static shouldShowLeftPaneBackdrop()
+	// {
+	// 	return this._isLeftPaneOverlayed && Screen.isMobile;
+	// }
 
-			this.backdrop.onCloseTrigger = () =>
-			{
-				// If they clicked the backdrop or something, close all overlayed panes.
-				this._isLeftPaneOverlayed = false;
-				this._isRightPaneOverlayed = false;
-				this.destroyBackdrop();
-			};
-		}
-		else if ( this.backdrop ) {
-			this.destroyBackdrop();
-		}
-	}
+	// private static checkBackdrop()
+	// {
+	// 	// // Ensure we have a backdrop if anything is overlayed.
+	// 	// // Otherwise ensure the backdrop is gone.
+	// 	// if ( this._isRightPaneOverlayed || this.shouldShowLeftPaneBackdrop() ) {
+	// 	// 	if ( this.backdrop ) {
+	// 	// 		return;
+	// 	// 	}
 
-	private destroyBackdrop()
-	{
-		this.backdrop.remove();
-		this.backdrop = undefined;
-	}
+	// 	// 	const Backdrop = getProvider<any>( 'Backdrop' );
+	// 	// 	this.backdrop = new Backdrop(
+	// 	// 		angular.element( document.getElementById( 'shell-body' ) as HTMLElement )
+	// 	// 	);
 
-	private bootstrap()
-	{
-		Api.sendRequest( '/web/library' )
-			.then( ( response: any ) =>
-			{
-				this.collections = this.collectionModel.populate( response.collections );
-				this.followedCollection = response.followedCollection
-					? new this.collectionModel( response.followedCollection )
-					: null;
-				this.developerCollection = response.developerCollection
-					? new this.collectionModel( response.developerCollection )
-					: null;
-				this.ownedCollection = response.ownedCollection
-					? new this.collectionModel( response.ownedCollection )
-					: null;
-				this.recommendedCollection = response.recommendedCollection
-					? new this.collectionModel( response.recommendedCollection )
-					: null;
-				this.bundleCollections = this.collectionModel.populate( response.bundleCollections );
+	// 	// 	this.backdrop.onCloseTrigger = () =>
+	// 	// 	{
+	// 	// 		// If they clicked the backdrop or something, close all overlayed panes.
+	// 	// 		this._isLeftPaneOverlayed = false;
+	// 	// 		this._isRightPaneOverlayed = false;
+	// 	// 		this.destroyBackdrop();
+	// 	// 	};
+	// 	// }
+	// 	// else if ( this.backdrop ) {
+	// 	// 	this.destroyBackdrop();
+	// 	// }
+	// }
 
-				this.isBootstrapped = true;
-				this.bootstrapPromiseResolve();
-			} );
+	// private static destroyBackdrop()
+	// {
+	// 	// this.backdrop.remove();
+	// 	// this.backdrop = undefined;
+	// }
 
-		this.broadcastModal.check();
-	}
+	// private static async bootstrap()
+	// {
+	// 	// getProvider<BroadcastModal>( 'BroadcastModal' ).check();
 
-	private clear()
-	{
-		this.collections = [];
-		this.followedCollection = undefined;
-		this.developerCollection = undefined;
-		this.ownedCollection = undefined;
-		this.recommendedCollection = undefined;
-		this.bundleCollections = [];
-	}
+	// 	const response = await Api.sendRequest( '/web/library' );
+
+	// 	this.collections = GameCollection.populate( response.collections );
+	// 	this.followedCollection = response.followedCollection
+	// 		? new GameCollection( response.followedCollection )
+	// 		: undefined;
+	// 	this.developerCollection = response.developerCollection
+	// 		? new GameCollection( response.developerCollection )
+	// 		: undefined;
+	// 	this.ownedCollection = response.ownedCollection
+	// 		? new GameCollection( response.ownedCollection )
+	// 		: undefined;
+	// 	this.recommendedCollection = response.recommendedCollection
+	// 		? new GameCollection( response.recommendedCollection )
+	// 		: undefined;
+	// 	this.bundleCollections = GameCollection.populate( response.bundleCollections );
+
+	// 	this.isBootstrapped = true;
+	// 	console.log( 'resolve the bootstrap' );
+	// 	bootstrapPromiseResolve();
+	// }
+
+	// private static clear()
+	// {
+	// 	this.collections = [];
+	// 	this.followedCollection = undefined;
+	// 	this.developerCollection = undefined;
+	// 	this.ownedCollection = undefined;
+	// 	this.recommendedCollection = undefined;
+	// 	this.bundleCollections = [];
+	// }
 }
