@@ -1,8 +1,8 @@
-angular.module( 'App.Views' ).config( function( $stateProvider )
-{
-	$stateProvider.state( 'discover.devlogs', {
-		url: '/devlogs',
-		abstract: true,
-		templateUrl: require( './devlogs.html' ),
-	} );
-} );
+// angular.module( 'App.Views' ).config( function( $stateProvider )
+// {
+// 	$stateProvider.state( 'discover.devlogs', {
+// 		url: '/devlogs',
+// 		abstract: true,
+// 		templateUrl: require( './devlogs.html' ),
+// 	} );
+// } );
