@@ -1,8 +1,16 @@
-import Overview from './overview/overview';
-import Games from './games/games';
+import Vue from 'vue';
+import { Component } from 'vue-property-decorator';
+import * as View from '!view!./devlogs.html';
 
-export default angular.module( 'App.Views.Devlogs', [
-	Overview,
-	Games,
-] )
-.name;
+import { AppPageHeader } from '../../../components/page-header/page-header';
+
+@View
+@Component({
+	name: 'route-discover-devlogs',
+	components: {
+		AppPageHeader,
+	},
+})
+export default class RouteDiscoverDevlogs extends Vue
+{
+}
