@@ -1,22 +1,22 @@
-angular.module( 'App.Views' ).config( function( $stateProvider )
-{
-	$stateProvider.state( 'discover.devlogs.games', {
-		abstract: true,
-		controller: 'Discover.Devlogs.GamesCtrl',
-		controllerAs: '$ctrl',
-		templateUrl: require( './games.html' ),
-		resolve: {
+// angular.module( 'App.Views' ).config( function( $stateProvider )
+// {
+// 	$stateProvider.state( 'discover.devlogs.games', {
+// 		abstract: true,
+// 		controller: 'Discover.Devlogs.GamesCtrl',
+// 		controllerAs: '$ctrl',
+// 		templateUrl: require( './games.html' ),
+// 		resolve: {
 
-			// We need translations loaded in for the filtering container, so we wait for "init".
-			filteringContainer: function( init, Game_Filtering_Container )
-			{
-				var filteringContainer = new Game_Filtering_Container();
+// 			// We need translations loaded in for the filtering container, so we wait for "init".
+// 			filteringContainer: function( init, Game_Filtering_Container )
+// 			{
+// 				var filteringContainer = new Game_Filtering_Container();
 
-				filteringContainer.isPersistent = false;
-				filteringContainer.shouldDetect = false;
+// 				filteringContainer.isPersistent = false;
+// 				filteringContainer.shouldDetect = false;
 
-				return filteringContainer;
-			}
-		}
-	} );
-} );
+// 				return filteringContainer;
+// 			}
+// 		}
+// 	} );
+// } );
