@@ -1,4 +1,4 @@
-angular.module( 'App.Views' ).controller( 'Library.OverviewCtrl', function( $scope, $timeout, App, GameCollection, Shell, gettextCatalog, payload )
+angular.module( 'App.Views' ).controller( 'Library.OverviewCtrl', function( $scope, $timeout, App, GameCollection, Shell, gettextCatalog )
 {
 	var _this = this;
 
@@ -27,31 +27,8 @@ angular.module( 'App.Views' ).controller( 'Library.OverviewCtrl', function( $sco
 		}
 	];
 
-	this.followedCollection = payload.followedCollection ? new GameCollection( payload.followedCollection ) : null;
-	this.developerCollection = payload.developerCollection ? new GameCollection( payload.developerCollection ) : null;
-	this.ownedCollection = payload.ownedCollection ? new GameCollection( payload.ownedCollection ) : null;
-	this.recommendedCollection = payload.recommendedCollection ? new GameCollection( payload.recommendedCollection ) : null;
-
-	this.bundleCollections = GameCollection.populate( payload.bundleCollections );
-
-	// Main collections that belong to the user get chunked out.
 	this.mainCollections = [];
-	if ( this.followedCollection ) {
-		this.mainCollections.push( this.followedCollection );
-	}
-
-	if ( this.developerCollection ) {
-		this.mainCollections.push( this.developerCollection );
-	}
-
-	if ( this.ownedCollection ) {
-		this.mainCollections.push( this.ownedCollection );
-	}
-
-	if ( this.recommendedCollection ) {
-		this.mainCollections.push( this.recommendedCollection );
-	}
-
+	this.bundleCollections = Shell.bundleCollections;
 	this.playlistCollections = [];
 	this.followedCollections = [];
 	$scope.$watchCollection( function()
@@ -60,6 +37,25 @@ angular.module( 'App.Views' ).controller( 'Library.OverviewCtrl', function( $sco
 	},
 	function( collections )
 	{
+		_this.mainCollections = [];
+		_this.bundleCollections = Shell.bundleCollections;
+
+		if ( Shell.followedCollection ) {
+			_this.mainCollections.push( Shell.followedCollection );
+		}
+
+		if ( Shell.developerCollection ) {
+			_this.mainCollections.push( Shell.developerCollection );
+		}
+
+		if ( Shell.ownedCollection ) {
+			_this.mainCollections.push( Shell.ownedCollection );
+		}
+
+		if ( Shell.recommendedCollection ) {
+			_this.mainCollections.push( Shell.recommendedCollection );
+		}
+
 		collections.forEach( function( collection )
 		{
 			if ( collection.type == 'playlist' && !collection.from_subscription ) {
