@@ -1,15 +1,15 @@
-angular.module( 'App.Views' ).config( function( $stateProvider )
-{
-	$stateProvider.state( 'search', {
-		abstract: true,
-		controller: 'SearchCtrl',
-		controllerAs: 'searchCtrl',
-		templateUrl: require( './search.html' ),
-		resolve: {
-			init: function( Translate )
-			{
-				return Translate.loadSection( 'main' );
-			},
-		}
-	} );
-} );
+// angular.module( 'App.Views' ).config( function( $stateProvider )
+// {
+// 	$stateProvider.state( 'search', {
+// 		abstract: true,
+// 		controller: 'SearchCtrl',
+// 		controllerAs: 'searchCtrl',
+// 		templateUrl: require( './search.html' ),
+// 		resolve: {
+// 			init: function( Translate )
+// 			{
+// 				return Translate.loadSection( 'main' );
+// 			},
+// 		}
+// 	} );
+// } );
