@@ -7,6 +7,7 @@ import { App } from './app-service';
 import { Analytics } from '../lib/gj-lib-client/components/analytics/analytics.service';
 import { Meta } from '../lib/gj-lib-client/components/meta/meta-service';
 import { Translate } from '../lib/gj-lib-client/components/translate/translate.service';
+import { History } from '../lib/gj-lib-client/components/history/history.service';
 
 import '../lib/gj-lib-client/components/error/error-module';
 import '../lib/gj-lib-client/components/body-classes/body-classes';
@@ -14,7 +15,6 @@ import '../lib/gj-lib-client/components/loading/loading';
 import '../lib/gj-lib-client/components/scroll/auto-scroll/auto-scroll';
 import '../lib/gj-lib-client/components/expand-when/expand-when';
 import '../lib/gj-lib-client/components/form/form';
-import '../lib/gj-lib-client/components/history/history';
 import '../lib/gj-lib-client/components/tooltip/tooltip.module';
 import '../lib/gj-lib-client/components/site-selector/site-selector';
 import '../lib/gj-lib-client/components/user/user-bar/user-bar';
@@ -45,7 +45,6 @@ export const AppModuleNg1 = angular.module( 'App', [
 	'gj.ExpandWhen',
 	'gj.Form',
 	'gj.Currency',
-	'gj.History',
 	'gj.Tooltip',
 	'gj.SiteSelector',
 	'gj.User.UserBar',
@@ -143,5 +142,6 @@ export const AppModuleNg1 = angular.module( 'App', [
 	Payload.initAngular( App, $transitions );
 	Analytics.initAngular( $rootScope );
 	Meta.initAngular( $rootScope );
+	History.initAngular( $rootScope );
 } )
 .name;
